Add automated integration tests
