+++ src/quillc/Cargo.toml
@@ -25,3 +25,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 toml = "0.5"
 tokio = { version = "1.7", features = ["macros", "rt-multi-thread"] }
+
+[dev-dependencies]
+test-case = "^1.1"
+tokio = { version = "1.7", features = ["macros", "rt-multi-thread", "time", "process"] }
