+++ src/quillc/tests/check_main_function/empty_project/README.md
@@ -1,0 +1,1 @@
+This test should fail since an empty project has no `main` source file.
