+++ src/quillc/tests/check_main_function/empty_project_with_main/main.ql
@@ -1,0 +1,1 @@
+// No `main` definition.
