+++ src/quillc/tests/check_main_function/empty_project_with_main/quill.toml
@@ -1,0 +1,1 @@
+name = "test"
