+++ src/quillc/tests/check_main_function/successful/README.md
@@ -1,0 +1,1 @@
+This test should succeed. It is the smallest possible valid program.
