+++ src/quillc/tests/check_main_function/successful/main.ql
@@ -1,0 +1,3 @@
+def main: Unit {
+    main = unit
+}
