+++ src/quillc/tests/check_main_function/successful/quill.toml
@@ -1,0 +1,1 @@
+name = "test"
