+++ src/quillc/tests/io_putchar/ascii/main.ql
@@ -1,0 +1,5 @@
+def putchar: Int -> Unit { compiler_intrinsic }
+
+def main: Unit {
+    main = putchar 65
+}
