+++ src/quillc/tests/io_putchar/ascii/output.txt
@@ -1,0 +1,1 @@
+A
